pyproject.toml
[project]
name = "ids-evaluation-framework"
version = "0.1.0"
description = "A framework for evaluating intrusion detection systems for IoT botnets."
authors = [
    { name = "Niklas Sandhu Peters", email = "niklas.s.peters@campus.tu-berlin.de" },
]
license-files = ["LICENSE"]
readme = { file = "README.txt", content-type = "text/markdown" }
requires-python = "~=3.13.0"

[project.scripts]
ids-eval = "ids_eval.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
core = [
    "adversarial-robustness-toolbox>=1.20.1",
    "catboost>=1.2.0",
    "codecarbon>=3.2.1",
    "colorlog>=6.10.1",
    "imbalanced-learn>=0.14.1",
    "lime>=0.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "psutil>=7.2.1",
    "pydantic-settings>=2.12.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "tensorflow>=2.20.0",
    "torch>=2.10.0",
    "typer[rich]>=0.21.1",
    "xgboost>=3.1.3",
#    "pyarrow>=23.0.0" # Currently disabled due to compatibility issues with tensorflow
]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "notebook>=7.5.3",
    "pandas-stubs>=2.3.3.260113",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "scipy-stubs>=1.17.0.2",
    "pandas-stubs~=3.0.0",
]

[tool.uv]
default-groups = ["dev", "core"]
package = true

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 120
